Change to pet policy
The prior revision noted the lack of policy in this area. If an affirmative policy is desired, I offer the proposed revision as more consistent with existing MIT & library norms, stated values, and policies; incorporating an appropriate balancing approach to reflect competing values; and promoting the goal of promoting resolution through collegial discussion.

Procedurally the text this revision replaces is a radical revision of the previous draft -- which had stood without public comment from its inception. And it lacks clear explanation of the new text. Our expectation should (and has been, I believe) that the revision process should be a transparent one, based on responses to public comments on the document, and explaining the rationale for revisions. So, I propose that we continue discussions in this document.